barber_backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==21.2.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
Flask==3.1.1
flask-cors==6.0.1
SQLAlchemy
psycopg2-binary
bcrypt==4.1.2
PyJWT==2.8.0
marshmallow==3.20.1
python-dotenv==1.0.0

